MR-77: Relocate the button
