--- src/Components/Trainers/Button/button.module.css
@@ -1,16 +1,0 @@
-.btnContainer {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 20px;
-  margin-right: 50px;
-}
-
-.btn {
-  padding: 10px 20px;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.btn:hover {
-  opacity: 0.8;
-}
